A time travelling debugger through multiple transitions in the state of the application that corresponds to each action in memory could be integrated with the router using the ngrx router-store package, adding the state of the router at the level of the state tree. Thus, displaying the right screen as we travel through the timeline.
The ngrx dev tools might run into performance issues because of the router state serialization which contains a lot of information. For that, configure a custom router state serializer that is only storing the url, the parameters of the url and the query parameters in the store state.

Now that the store state contains information about the router, we need to add a reducer for the router state at the level of reducers to handle that state.