For implementing a ngrx feature, start by writing the actions; create a new actions file and define in it an enum containing action types. Each action has an origin and event.
The first action is courseRequested, dispatched by the origin of the action [View Course Page] in order to report back to the store that we are looking for a given course that we did not find inside the store. The event that is linked to this action is Course Requested. In this case we are triggering an effect that is going to call the backend to retrieve the course with the given course ID. Once that happens, that effect is going to dispatch a new action, this time around the CourseLoaded action. This means that the origin of action is no longer the [View Course Page], the origin of action is the [Courses API] call that we just made. The event that is linked to this action is Course Loaded.

Next, define a class for each action implementing the Action interface and define the read-only property type by pointing to the corresponding element of the enum. Define a constructor for each class. In case of courseRequested constructor, it is going to take a public property called payload containing a property courseId which is of type number that identifies what course was requested. In case of CourseLoaded constructor,the property of the payload is the course itself, the Course data is going to be dispatched to the store and via a reducer, the course is going to be saved in the in-memory store.

Finally, define a union type called courseActions. So, an element of type courseActions is going to be either of type courseRequested or CourseLoaded.

Now that we have defined the actions of a new ngrx feature, go ahead and define the form of the corresponding courses state in the store using ngrx entity. Thus, the Collection of courses (entities) in the store.